requirements.txt
customtkinter
mysql.connector